normalized heuristics

+++ l1/code/heuristic.h
@@ -1,5 +1,5 @@
 #pragma once
 
-int heuristic1(Puzzle& puzzle);
-int heuristic2(Puzzle& puzzle);
-int heuristic3(Puzzle& puzzle);
+float heuristic1(Puzzle& puzzle);
+float heuristic2(Puzzle& puzzle);
+float heuristic3(Puzzle& puzzle);
